requirements/dev.txt
-r base.txt

#devlibs
django-debug-toolbar==1.6
django-model-utils==2.6.1

#tinymce
django-tinymce
